Move use cases to data modules but leave tests in feature modules
