+++ backend/search/src/jsTest/kotlin/com/gchristov/thecodinglove/search/usecase/RealPreloadSearchResultUseCaseTest.kt
@@ -5,6 +5,7 @@ import com.gchristov.thecodinglove.kmpcommontest.FakeCoroutineDispatcher
 import com.gchristov.thecodinglove.searchdata.model.SearchError
 import com.gchristov.thecodinglove.searchdata.model.SearchSession
 import com.gchristov.thecodinglove.searchdata.usecase.PreloadSearchResultUseCase
+import com.gchristov.thecodinglove.searchdata.usecase.RealPreloadSearchResultUseCase
 import com.gchristov.thecodinglove.searchdata.usecase.SearchWithHistoryUseCase
 import com.gchristov.thecodinglove.searchtestfixtures.*
 import kotlinx.coroutines.ExperimentalCoroutinesApi
