+++ backend/search/src/jsTest/kotlin/com/gchristov/thecodinglove/search/usecase/RealSearchWithSessionUseCaseTest.kt
@@ -4,6 +4,7 @@ import arrow.core.Either
 import com.gchristov.thecodinglove.kmpcommontest.FakeCoroutineDispatcher
 import com.gchristov.thecodinglove.searchdata.model.SearchError
 import com.gchristov.thecodinglove.searchdata.model.SearchSession
+import com.gchristov.thecodinglove.searchdata.usecase.RealSearchWithSessionUseCase
 import com.gchristov.thecodinglove.searchdata.usecase.SearchWithHistoryUseCase
 import com.gchristov.thecodinglove.searchdata.usecase.SearchWithSessionUseCase
 import com.gchristov.thecodinglove.searchtestfixtures.*
