+++ backend/slack/src/jsTest/kotlin/com/gchristov/thecodinglove/slack/RealVerifySlackRequestUseCaseTest.kt
@@ -3,6 +3,7 @@ package com.gchristov.thecodinglove.slack
 import arrow.core.Either
 import com.gchristov.thecodinglove.kmpcommontest.FakeCoroutineDispatcher
 import com.gchristov.thecodinglove.kmpcommontest.FakeLogger
+import com.gchristov.thecodinglove.slackdata.RealVerifySlackRequestUseCase
 import com.gchristov.thecodinglove.slackdata.VerifySlackRequestUseCase
 import com.gchristov.thecodinglove.slackdata.domain.SlackConfig
 import com.gchristov.thecodinglove.slacktestfixtures.FakeSlackApiRequest
